[ADD] point_of_sale patch
================================================================================

This commit filters or validates the mapping result in order to take care of the
undifined returns. As it is construed as a mapping function, we cannot take care
within the function itself but must handle it in downstream code.

Alternatively a wrapper could be made to encapuslate this safety check as well.

================================================================================
## To make a commit, type your commit message and press CTRL-ENTER. To cancel
## the commit, close the window. To sign off on the commit, press CTRL-S.
##
## You may also reference or close a GitHub issue with this commit.  To do so,
## type `#` followed by the `tab` key.  You will be shown a list of issues
## related to the current repo.  You may also type `owner/repo#` plus the `tab`
## key to reference an issue in a different GitHub repo.

 ...der-to-take-care-of-the-undifined-returns.patch | 25 ++++++++++++++++++++++
 1 file changed, 25 insertions(+)